Address review comments from davidradl
- Rename LOCK to CACHE_LOCK for better clarity
- Move reference count increment before LOG.debug to avoid side effects in logging
- Move ConnectionPoolConfig to its own file for better code organization
- Add comment explaining why maxRequestsPerHost uses maxTotalConnections
- Clarify QPS acronym (Queries Per Second) in documentation
- Clarify 'route' means unique host:port combinations
- Improve connection-timeout description to emphasize it must be less than overall timeout
